# openapi.yaml
openapi: 3.0.0
info:
  title: 'GamerVault Pro: Real-Time PC Gaming Deals & Freebies'
  description: >-
    The definitive API for tracking deals and free games in the PC Ecosystem (Steam and Epic Games Store). 
    Provides high-fidelity data with 100% authentic regional pricing mapped directly from the Official Steam API and Epic Storefronts.
  version: 1.0.0
servers:
  - url: 'https://gamervault-pro.onrender.com/api/v1'
    description: Production Server (Render)
paths:
  /deals/freebies:
    get:
      tags:
        - Freebies
      summary: Get currently 100% Free PC Games
      description: Captures active giveaways from the Epic Games Store and "Free to Keep" promotions on Steam.
      operationId: getFreebies
      responses:
        '200':
          description: List of free games found.
          content:
            application/json:
              example:
                - title: "Clone Drone in the Danger Zone"
                  store: "Epic Games Store"
                  url: "https://www.epicgames.com/store/p/..."
                  store_url: "https://store.steampowered.com/app/1"
                  status: "Currently Free"
                  image: "https://cdn1.epicgames.com/..."

  /deals/latest:
    get:
      tags:
        - Deals
      summary: List trending PC deals (Steam focus)
      description: Aggregated list containing the daily best savings from Steam and Epic Games, including regional price injection.
      operationId: getLatestDeals
      parameters:
        - name: store
          in: query
          description: Filter by store name (e.g., Steam, Epic).
          schema:
            type: string
        - name: currency
          in: query
          description: Primary regional currency (Available for PRO users).
          schema:
            type: string
            enum: [USD, BRL, TRY, ARS]
            default: USD
        - name: convert_to
          in: query
          description: Live FX conversion factor. Converts the regional price to another fiat currency using real-time rates.
          schema:
            type: string
            enum: [USD, EUR, GBP, CAD, BRL]
      responses:
        '200':
          description: Real-time deals list processed.
          content:
            application/json:
              example:
                - title: "Cyberpunk 2077"
                  store: "Steam"
                  steam_app_id: "1091500"
                  price:
                    sale_price_usd: 29.99
                    normal_price_usd: 59.99
                    savings: "50%"
                    regional:
                      BRL: 149.95
                      TRY: 959.68
                      ARS: 29990.00
                      converted_to_EUR: 28.50
                  url: "https://store.steampowered.com/app/1091500"

  /deals/search:
    get:
      tags:
        - Search
      summary: Live Search (Premium API)
      description: >-
        Perform a Deep Live Search against a database of +50,000 PC Games. 
        Instead of returning only current trending deals, this feature hits the market in real-time to find exact matches for your query (e.g. 'Elden Ring', 'GTA'). 
        Ideal for Discord Bots and targeted price tracking. Recommended for PRO tiers.
      operationId: searchDeals
      parameters:
        - name: q
          in: query
          required: true
          description: Game title to search for (e.g., 'Cyberpunk', 'Elden Ring').
          schema:
            type: string
      responses:
        '200':
          description: Search results.
          content:
            application/json:
              example:
                - title: "Elden Ring"
                  store: "Steam"
                  price: 59.99
                  url: "https://..."

components:
  securitySchemes:
    RapidAPIKey:
      type: apiKey
      name: X-RapidAPI-Key
      in: header
    ProxySecret:
      type: apiKey
      name: X-RapidAPI-Proxy-Secret
      in: header